Renamed `feedSourceTypes.types.ts` to `feedTypes.types.ts`
