+++ packages/pwa/src/components/views/View.tsx
@@ -13,8 +13,8 @@ import {Views} from '@shared/lib/views.shared';
 import {AsyncStatus} from '@shared/types/asyncState.types';
 import {FeedItemContentType} from '@shared/types/feedItemContent.types';
 import type {FeedItem, FeedItemId} from '@shared/types/feedItems.types';
-import type {FeedType} from '@shared/types/feedSourceTypes.types';
 import type {FeedSubscriptionId} from '@shared/types/feedSubscriptions.types';
+import type {FeedType} from '@shared/types/feedTypes.types';
 import type {TagId} from '@shared/types/tags.types';
 import type {Supplier} from '@shared/types/utils.types';
 import type {ViewGroupByOption, ViewSortByOption, ViewType} from '@shared/types/views.types';
