+++ packages/shared/src/lib/eventLog.shared.ts
@@ -19,8 +19,8 @@ import type {
 import {EventType} from '@shared/types/eventLog.types';
 import type {ExperimentId, ExperimentType} from '@shared/types/experiments.types';
 import type {FeedItemActionType, FeedItemId} from '@shared/types/feedItems.types';
-import type {FeedType} from '@shared/types/feedSourceTypes.types';
 import type {FeedSubscriptionId} from '@shared/types/feedSubscriptions.types';
+import type {FeedType} from '@shared/types/feedTypes.types';
 import type {ThemePreference} from '@shared/types/theme.types';
 
 /**
