+++ packages/shared/src/storage/feedSubscriptions.storage.ts
@@ -12,7 +12,6 @@ import {parseYouTubeChannelId} from '@shared/parsers/youtube.parser';
 
 import type {AccountId} from '@shared/types/accounts.types';
 import type {DeliverySchedule} from '@shared/types/deliverySchedules.types';
-import {FeedType} from '@shared/types/feedSourceTypes.types';
 import {FeedSubscriptionActivityStatus} from '@shared/types/feedSubscriptions.types';
 import type {
   FeedSubscription,
@@ -22,6 +21,7 @@ import type {
   RssFeedSubscription,
   YouTubeChannelFeedSubscription,
 } from '@shared/types/feedSubscriptions.types';
+import {FeedType} from '@shared/types/feedTypes.types';
 import type {Result} from '@shared/types/results.types';
 
 import type {
