+++ packages/shared/src/types/eventLog.types.ts
@@ -3,8 +3,8 @@ import type {Actor} from '@shared/types/actors.types';
 import type {Environment} from '@shared/types/environment.types';
 import type {ExperimentId, ExperimentType} from '@shared/types/experiments.types';
 import type {FeedItemActionType, FeedItemId} from '@shared/types/feedItems.types';
-import type {FeedType} from '@shared/types/feedSourceTypes.types';
 import type {FeedSubscriptionId} from '@shared/types/feedSubscriptions.types';
+import type {FeedType} from '@shared/types/feedTypes.types';
 import type {ThemePreference} from '@shared/types/theme.types';
 import type {BaseStoreItem} from '@shared/types/utils.types';
